scripts/async/requirements.txt
datasets
openai
python-dotenv
